clean up select storybook
