+++ packages/storybook/stories/select/disabled.js
@@ -1,0 +1,70 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+import { boolean } from '@kadira/storybook-addon-knobs';
+
+import SourceViewer from '../../components/source_viewer';
+
+module.exports = function disabled() {
+  const props = {
+    disabled: boolean('Disabled', true),
+  };
+
+  const sources = [
+    {
+      label: 'React',
+      source: `
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Select} from '@orion-ui/react/lib/2016-12-01';
+
+class App extends React.Component {
+render() {
+  const options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  return (
+    <Select options={options} disabled={${props.disabled}} />
+  )
+}
+}
+
+ReactDOM.render(React.createElement(App), document.body);
+      `,
+    },
+    {
+      label: 'Angular 1.5.x',
+      source: `
+// app controller
+import 'angular';
+
+angular.module('app', [])
+.controller('AppController', function() {
+  var app = this;
+  app.options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  app.disabled = ${props.disabled};
+}]);
+
+// app.html
+
+<!doctype html>
+<html lang="en" ng-app="app">
+<body ng-controller="AppController as app">
+  <orion-select options="{{app.options}}" disabled="{{app.disabled}}" />
+</body>
+</html>
+      `,
+    },
+  ];
+
+  return (
+    <div>
+      <SourceViewer sources={sources} />
+    </div>
+  );
+};
