+++ packages/storybook/stories/select/no-search-results.js
@@ -1,0 +1,74 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+import { boolean, text } from '@kadira/storybook-addon-knobs';
+
+import SourceViewer from '../../components/source_viewer';
+
+module.exports = function noSearchResults() {
+  const props = {
+    searchable: boolean('Searchable', true),
+    // query text makes the select open
+    query: text('Query', 'Hello World'),
+  };
+
+  const sources = [
+    {
+      label: 'React',
+      source: `
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Select} from '@orion-ui/react/lib/2016-12-01';
+
+class App extends React.Component {
+render() {
+  const options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  return (
+    <Select options={options} searchable={${props.searchable}} query="${props.query}" />
+  )
+}
+}
+
+ReactDOM.render(React.createElement(App), document.body);
+      `,
+    },
+    {
+      label: 'Angular 1.5.x',
+      source: `
+// app controller
+import 'angular';
+
+angular.module('app', [])
+.controller('AppController', function() {
+  var app = this;
+  app.options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  app.open = ${props.open};
+  app.searchable = ${props.searchable};
+  app.query = "${props.query}";
+}]);
+
+// app.html
+
+<!doctype html>
+<html lang="en" ng-app="app">
+<body ng-controller="AppController as app">
+  <orion-select options="{{app.options}}" searchable="{{app.searchable}}" query="{{app.query}}" />
+</body>
+</html>
+      `,
+    },
+  ];
+
+  return (
+    <div>
+      <SourceViewer sources={sources} />
+    </div>
+  );
+};
