+++ packages/storybook/stories/select/option-focus.js
@@ -1,0 +1,75 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+import { number } from '@kadira/storybook-addon-knobs';
+
+import SourceViewer from '../../components/source_viewer';
+
+module.exports = function optionFocus() {
+  const props = {
+    focusIndex: number('Focus Index', 0, {
+      range: true,
+      min: 0,
+      max: 1,
+      step: 1,
+    }),
+  };
+
+  const sources = [
+    {
+      label: 'React',
+      source: `
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Select} from '@orion-ui/react/lib/2016-12-01';
+
+class App extends React.Component {
+render() {
+  const options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  return (
+    <Select options={options} open={true} focusIndex={${props.focusIndex}} />
+  )
+}
+}
+
+ReactDOM.render(React.createElement(App), document.body);
+      `,
+    },
+    {
+      label: 'Angular 1.5.x',
+      source: `
+// app controller
+import 'angular';
+
+angular.module('app', [])
+.controller('AppController', function() {
+  var app = this;
+  app.options = [
+    { value: 'one', label: 'One' },
+    { value: 'two', label: 'Two' }
+  ];
+
+  app.focusIndex = ${props.focusIndex};
+}]);
+
+// app.html
+
+<!doctype html>
+<html lang="en" ng-app="app">
+<body ng-controller="AppController as app">
+  <orion-select options="{{app.options}}" focusIndex="{{app.focusIndex}}" />
+</body>
+</html>
+      `,
+    },
+  ];
+
+  return (
+    <div>
+      <SourceViewer sources={sources} />
+    </div>
+  );
+};
